The dropdown menu setup for the hover event in CSS3 is done.
The page is primarily finished if I decide not to spend more time building the other sidebar pages.

I may want to add CSS or JS tooltips to indicate what sections are finished and ready to go.

At this point I want to code some Javascript to give the dropdown menu a use on mobile devices (where hover doesn't work).