Changed the CSS hexagon shape code for the Highlights, and recent posts in Footer .date to use ::before or ::after depending on the situation.
NOTE:  The Font Awesome icons used in Highlights use ::before so if you try to use ::before with the downward pointing arrows you will over-write the icons in the code and they disappear.  In that case I had to use ::after.

Theoretically though it doesn't matter if you use ::after or ::before when you use positioning absolute.